% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicCalc.R
\name{add_num}
\alias{add_num}
\title{Add Two Numbers}
\usage{
add_num(x, y)
}
\arguments{
\item{x}{A numeric value.}

\item{y}{A numeric value.}
}
\value{
A numeric value representing \code{x + y}.
}
\description{
Computes the sum of two numeric values.
}
\examples{
add_num(3, 5)

}
