% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicCalc.R
\name{calc_av}
\alias{calc_av}
\title{Average of a Numeric Vector}
\usage{
calc_av(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
Numeric mean.
}
\description{
Computes the arithmetic mean of a numeric vector with input checks.
}
\examples{
calc_av(c(10, 20, 30, 40))

}
