% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interestCalc.R
\name{cpond_int}
\alias{cpond_int}
\title{Compound Interest Calculation}
\usage{
cpond_int(principal, rate, n, time)
}
\arguments{
\item{principal}{Numeric principal amount \eqn{P}.}

\item{rate}{Annual interest rate \eqn{r} (as a decimal).}

\item{n}{Number of compounding periods per year \eqn{n}.}

\item{time}{Time in years \eqn{t}.}
}
\value{
Numeric value of the final amount \eqn{A}.
}
\description{
Computes compound interest using \eqn{A = P (1 + r/n)^{n t}}.
}
\examples{
cpond_int(1000, 0.05, 4, 3)

}
