% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicCalc.R
\name{dist_2d}
\alias{dist_2d}
\title{Calculate Distance Between Two Points}
\usage{
dist_2d(x1, x2, y1, y2)
}
\arguments{
\item{x1, x2}{Numeric x-coordinates.}

\item{y1, y2}{Numeric y-coordinates.}
}
\value{
Numeric distance.
}
\description{
Computes Euclidean distance between points \eqn{(x1, y1)} and \eqn{(x2, y2)}.
}
\examples{
dist_2d(0, 3, 0, 4)  # returns 5

}
