% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicCalc.R
\name{div_num}
\alias{div_num}
\title{Divide Two Numbers Safely}
\usage{
div_num(numerator, denominator)
}
\arguments{
\item{numerator}{Numeric.}

\item{denominator}{Numeric.}
}
\value{
Numeric quotient.
}
\description{
Performs division with a check for division by zero.
}
\examples{
div_num(10, 2)

# This will throw an error:
# div_num(5, 0)

}
