% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicCalc.R
\name{mult_num}
\alias{mult_num}
\title{Multiply Two Numbers}
\usage{
mult_num(x, y)
}
\arguments{
\item{x}{Numeric.}

\item{y}{Numeric.}
}
\value{
Numeric product.
}
\description{
Computes the product of two numeric inputs.
}
\examples{
mult_num(4, 6)

}
