% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentCalc.R
\name{perc_chg}
\alias{perc_chg}
\title{Calculate Percent Change}
\usage{
perc_chg(old, new)
}
\arguments{
\item{old}{A numeric value (or vector) representing the original amount.}

\item{new}{A numeric value (or vector) representing the new amount.}
}
\value{
A numeric value (or vector) of percent changes.
}
\description{
This function computes the percent change from an old value to a new value,
using the formula \code{(new - old) / old * 100}.
}
\examples{
perc_chg(100, 120)   # 20\% increase
perc_chg(c(50, 80), c(75, 40))

}
