% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interestCalc.R
\name{simp_int}
\alias{simp_int}
\title{Simple Interest Calculation}
\usage{
simp_int(principal, rate, time)
}
\arguments{
\item{principal}{Numeric principal amount \eqn{P}.}

\item{rate}{Annual interest rate \eqn{r} (as a decimal).}

\item{time}{Time in years \eqn{t}.}
}
\value{
Numeric value of the total interest.
}
\description{
Computes simple interest using \eqn{I = P \cdot r \cdot t}.
}
\examples{
simp_int(1000, 0.05, 3)

}
