% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algLinCalc.R
\name{solv_lin}
\alias{solv_lin}
\title{Solve a Simple Linear Equation}
\usage{
solv_lin(a, b)
}
\arguments{
\item{a}{Numeric coefficient for \eqn{x}.}

\item{b}{Numeric intercept term.}
}
\value{
A single numeric value giving the solution for \eqn{x}.
}
\description{
Solves the equation \eqn{a x + b = 0} for \eqn{x}, assuming \eqn{a != 0}.
}
\examples{
solv_lin(2, -4)  # x = 2

}
