% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicCalc.R
\name{sub_num}
\alias{sub_num}
\title{Subtract Two Numbers}
\usage{
sub_num(x, y)
}
\arguments{
\item{x}{A numeric value.}

\item{y}{A numeric value.}
}
\value{
Numeric difference of \code{x - y}.
}
\description{
Computes the difference \code{x - y}.
}
\examples{
sub_num(10, 3)

}
